Move templatetags from prices_i18n to prices
