+++ tests/test_templatetags.py
@@ -3,7 +3,7 @@ import functools
 
 import pytest
 from django.utils import translation
-from django_prices.templatetags import prices, prices_i18n
+from django_prices.templatetags import prices
 
 from prices import Money, TaxedMoney, percentage_discount
 
@@ -31,7 +31,7 @@ def test_non_existing_locale(money_fixture):
     #     Babel needs to support Fuzzy Locales
     #     https://github.com/python-babel/babel/issues/30
     translation.activate("zz")
-    amount = prices_i18n.amount(money_fixture, format="html")
+    amount = prices.amount(money_fixture, format="html")
     assert amount  # No exception, success!
 
 
@@ -48,45 +48,40 @@ def test_non_cannonical_locale_zh_CN(money_fixture, settings):
     settings.LANGUAGE_CODE = "en_US"
 
     # Checking format of the default locale
-    amount = prices_i18n.amount(money_fixture, format="html")
+    amount = prices.amount(money_fixture, format="html")
     assert amount == '<span class="currency">$</span>10.00'
 
     # Checking if 'zh_CN' has changed the format
     translation.activate("zh_CN")
-    amount = prices_i18n.amount(money_fixture, format="html")
+    amount = prices.amount(money_fixture, format="html")
     assert amount == '<span class="currency">US$</span>10.00'  # 'US' before '$'
 
 
 def test_templatetag_amount(money_fixture):
     amount = prices.amount(money_fixture)
-    assert amount == '10 <span class="currency">USD</span>'
-
-
-def test_templatetag_i18n_amount(money_fixture):
-    amount = prices_i18n.amount(money_fixture)
     assert amount == "$10.00"
 
 
-def test_templatetag_i18n_amount_html(money_fixture):
-    amount = prices_i18n.amount(money_fixture, format="html")
+def test_templatetag_amount_html(money_fixture):
+    amount = prices.amount(money_fixture, format="html")
     assert amount == '<span class="currency">$</span>10.00'
 
 
-def test_templatetag_i18n_amount_wrong_param(money_fixture):
-    amount = prices_i18n.amount(money_fixture, format="test")
+def test_templatetag_amount_wrong_param(money_fixture):
+    amount = prices.amount(money_fixture, format="test")
     assert amount == "$10.00"
 
 
 def test_get_currency_fraction_USD():
-    result = prices_i18n.get_currency_fraction("USD")
+    result = prices.get_currency_fraction("USD")
     assert result == 2
 
 
 def test_get_currency_fraction_unknown_currency():
-    result = prices_i18n.get_currency_fraction(("test"))
+    result = prices.get_currency_fraction(("test"))
     assert result == 2
 
 
 def test_format_price_invalid_value():
-    result = prices_i18n.format_price("invalid", "USD")
+    result = prices.format_price("invalid", "USD")
     assert result == ""
